refactor: Moved dataset async loader util, added more unit tests for time slice
